pyproject.toml
[project]
name = "langchainclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.43.0",
    "fastapi>=0.115.6",
    "google-generativeai>=0.8.3",
    "ipywidgets>=8.1.5",
    "langchain-anthropic>=0.3.1",
    "langchain-community>=0.3.14",
    "langchain-google-genai>=2.0.9",
    "polars>=1.19.0",
    "pyarrow>=19.0.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "typer>=0.15.1",
    "widgetsnbextension>=4.0.13",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.0.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["langchainclassifier"]
